examples-requirements.txt
aiohttp>=3.10.3
numpy>=2.0.1
transformers>=4.44.0
torch>=2.4.0
rich>=13.7.1
av>=12.3
opencv-python>=4.10
pillow>=10.4.0
numexpr>=2.10.1
pydantic>=2.8.2
python-dotenv>=1.0.1
scipy>=1.14.0
matplotlib>=3.9.1